requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
supabase
pymupdf
zai
fitz